+++ src/DefaultLogger.h
@@ -7,7 +7,7 @@
 /// @brief Default logger class used by the ThingsBoard class to log messages into the console
 class DefaultLogger : public ILogger {
   public:
-    virtual int logf(const char *format, ...) const override;
+    virtual int log(const char *format, ...) const override;
 };
 
 #endif // Default_Logger_h
